+++ index.js
@@ -57,45 +57,9 @@ app.get('/auth/google/failure',(req,res)=>{
     res.send('Failure from Google sign in');
   })
 
-//GET AND POST SIGNUP
-app.post('/signup',[check('username').isEmail().withMessage('please enter valid email'),
-check('password','password length should be greater than 5 ').isLength({min:5}),
-check('confirmPassword').custom((value,{req})=>{
-    if(value != req.body.password){
-        throw new Error('Password Have To Match');
-    }
-    return true
-})]
-,async(req,res)=>{
-    const displayName = req.body.displayName;
-    const username = req.body.username.toLowerCase();
-    const password = req.body.password; 
-    const user = await User.findOne({username:username})
-    if(user)return res.status(400).render('signup.ejs',{user:req.user,errorMessage:"User Already Exists",oldInput:{displayName:req.body.displayName,username:req.body.username,password:req.body.password,confirmPassword:req.body.confirmPassword},validationErrors:[]}); 
-    const errors = validationResult(req);
-    if(!errors.isEmpty()){
-        console.log(errors.array());
-        return res.status(422).render('signup.ejs',
-        {user:req.user,
-        errorMessage:errors.array()[0].msg,
-        oldInput:{displayName:req.body.displayName,username:req.body.username,password:req.body.password,confirmPassword:req.body.confirmPassword},
-        validationErrors:errors.array()})
-    }
-    let newpass= await bcrypt.hash(req.body.password,10);
-    const newUser = new User({displayName,username,password});
-    newUser.save();
-    console.log(newUser);
-    // await User.create({displayName,username,newpass});
-    res.redirect('/login');
- })
 
-app.get('/signup',(req,res)=>{
-  
-    res.render('signup.ejs',{user:req.user,
-        errorMessage:null,
-        oldInput:{displayName:"",username:"",password:"",confirmPassword:""},validationErrors:[]}
-        ); 
-});
+
+
